Fix tag names in XMLToBinary converter
If applied, this commit will fix the command stream XML rendering issues
in the browser. Prior to this commit, some of the tags in the XML
structure weren't following the XML naming rule that an element name
cannot start with digits. Due to this, the command stream XML wasn't
being rendered properly in the browser.

Signed-off-by: Shubham <shubham@arm.com>