Fix unit test Linux kernel header build dependency
When SCons is building unit tests with dependencies to the Linux
kernel's user space headers, the build sometimes fails because there is
a race condition between the generation of the kernel headers and
building the unit tests.

The race condition occurs because when SCons parses the sources files
for dependencies, it will ignore dependencies to header files that
doesn't exist at the time of parsing. This means SCons is unaware of the
dependencies to the kernel headers and will start building the object
files for the unit tests before the kernel headers have been fully
generated.

To resolve this, the object files for the unit tests now have a direct
dependency on the Linux kernel's user space headers to ensure that they
are always generated before building the unit tests.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>