Add stripe config override settings
The environment variable ETHOSN_SUPPORT_LIBRARY_DEBUG_STRIPE_CONFIG can
optionally be set to a simple configuration file which enables/disables
stripe configuration settings, for example to force a specific strategy
to be used. This is useful for debugging but is not a production
feature.

StripeSplitInfo has been renamed and expanded into StripeConfig, which
contains all the settings that can be overriden in the config file.
StripeGenerator now loops over block configs internally, rather than
requiring the calling code to do it. This also simplifies the
GenerateStripes interface.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>