Improve command stream debugging
Add comments to indicate the agent IDs, which agent a dependency points
to and what kind of agent a dependency points to.

Remove some invalid dependencies from the test command stream. This is
the first agent in the command stream, and so it's not possible to have
any read dependencies (read dependencies always go backwards, and there
is nothing to go backwards to!). Previously this didn't matter as the
Xml<>Binary conversion just did a dumb conversion, but with the
improvements in this patch, it 'resolves' the dependencies to add
helpful comments, and so the dependencies need to be valid otherwise it
will crash. (Same but in reverse for write/schedule dependencies).

Signed-off-by: Rob Hughes <robert.hughes@arm.com>