Fix calculation of glue DRAM buffer sizes
Previously the code was calculating the buffer size assuming that all
buffers were NHWCB, which is not always the case.

Merge two of the existing CalculateBufferSize functions into the more
generic Utils.hpp and use this instead.

Prevent ConstantParts being created for parts that are not connected
to anything (e.g. weights/biases), as it is no longer possible to create
a ConstantPart for weights. This simplifies the GraphOfParts. These
parts could never participate in the proper network anyway as they
would never be connected to an input or output.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>