Honour firmware limits on stripe ratios
Limit the number of MCE stripes per PLE stripe and IFM+WGT stripes per
PLE stripe.

The firmware has a limitation on the maximum number of MCE stripes that
can be consumed by a PLE stripe. If this limit is broken then the
firmware will get stuck in a deadlock. This patch exposes this
limitation of the firmware in the capabilities struct, and uses it in
the support library to prevent generating plans that would break it.

There is also a related but distinct limitation on the maximum number of
IFM and weight stripes that need to be scheduled before a PLE stripe
can be scheduled, which is now also handled in the same way.

The PLE input stripe was not being set correctly, which has been fixed
so that it can be used to calculate if the plan is acceptable for the
firmware.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>