Fix for MceS/PleS stripes out of order, part 2
(The return of the PLE)

We need to ensure that PLE stripes are scheduled in the same order as
the MCE stripes, otherwise the firmware will deadlock. This can happen
in a strategy 0 cascade if an MCE stripe is scheduled but the following
PLE stripe is not, because there is no space in the queue. An MCE and
PLE stripe from the preceding layer can then get scheduled, and this
means that we missed the PLE stripe from the second layer. To prevent
this, we make sure that a PLE stripe is not scheduled until the next
PLE stripe in the next layer is needed, so that the order is correct.
This is done by adding a schedule dependency on the following PLE
agent.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>