Fix for MceS/PleS stripes out of order
We need to ensure that MCE stripes are scheduled in the same order as
the PLE stripes, otherwise the firmware will deadlock. This can happen
in a strategy 0 cascade if an MCE stripe is scheduled but the following
PLE stripe is not, because it is not yet needed. An MCE and PLE stripe
from the following layer can then get scheduled, and this means that we
missed the PLE stripe from the first layer. To prevent this, we make
sure that an MCE stripe is not scheduled unless the PLE stripe
following it is needed, so that it will be scheduled before any other
PLE stripes. This is done by adding a schedule dependency on the
following Mce agent, so that the MCE stripe from the first layer will
not be scheduled until the MCE stripe from the second layer is
scheduled.

Increase the max number of schedule dependencies to 2, as we need an
extra slot for the new dependency. Improve the way the dependency
generation code chooses which dependency slot to use (find the first
free slot, rather than hardcoding to a particular slot).

Signed-off-by: Rob Hughes <robert.hughes@arm.com>