Fix stripe shapes for pooling
The support library was allowing PLE output stripe shapes that weren't
multiples of the brick group width/height for pooling, which isn't
valid.

This patch makes two changes to fix this:

1. Lonely and beginning plans that have a pooling in them will use
multiple blocks per stripe as necessary so that the PLE output stripe
is at least 8.

2. Continue and end plans will not be generated if they would result in
an invalid stripe shape from the PLE.

Also simplify and fix the calculation of PLE output stripe shapes.

There is a future potential for double-buffering stripes for the PLE
to lift these limitations, but it was found that this caused issues
with the dependencies and tile size calculations and so this is left
as a future performance improvement.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>