Set Operation Ids for Input and Output Buffers
CascadingCommandStreamGeneration needs to add the correct operation ids
of the input and output buffers for the BufferManager. This change adds
a new optional field in the Buffer class to store that information.
Also the Input and Output parts has been updated to store the operation
ids to the newly added field.

Signed-off-by: Mohamed Elzahhar <Mohamed.Elzahhar@arm.com>