Fix supertensor shapes with reshape
The supertensor shape in the command stream was being set from the DRAM
buffer, which might be different to the SRAM buffer shape if there
was a reshape. This can be a problem as the stripe shape and supertensor
shape might not be compatible.

This patch changes to use the SRAM buffer shape instead in the case of
reshapes.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>