Add Support for Split in CascadingCommandStreamGenerator
 - Support for Split op implemented
 - Unit tests implemented

There was an issue with split ops, with split ops being adjacent, the
split ops would be done immediately after each other. This means if
there was a large number of split ops the dependencies on the following
agents may overflow the window.

Sort the ops as a preprocess step so now each full branch should be
performed before the next split op.

Change how split ops store its information. Store the full offset
within the source tensor. This makes the dram buffer calculation easier
when generating the command stream.

Add functions to calculate the Dram offset for NHWC and NHWCB tensors.

Fix a bug with branching and disjoint subgraphs. There wasn't a
dependency to prevent the branching/disjoint operations from starting
early. Add a dependency between OFM and IFM streamers.

Change Concat and Split Ops to inherit from DmaOp. (Not sure about
this), it makes the logic for the weight streamer simpler although that
might be able to be simplified to be the same logic as the OFM and IFM
streamers.

Signed-off-by: Matthew Haddon <Matthew.Haddon@arm.com>