Fix cascading intermediate buffer lifetimes
The previous calculation was not accounting for the fact that Ops
execute in parallel with each other and so it ended up allocating all
buffers at address zero, and they overwrote each other.

This new patch walks the graph in either direction until it finds
another DRAM buffer, at which point it knows the extents of the
lifetime.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>