Input depth splitting strategies with multiplier
Because the firmware has limits to the maximum number of MCE stripes per
PLE stripe, it means that input depth splitting strategies may result
in command streams that are not supported by the firmware.

This patch fixes the issue by increasing the number of (lonely) plans
which split input depth, by having a multiplier for the input depth.
This will result in less MCE stripes per PLE stripe.

This fixes estimation for several networks that exposed this issue.

In order to not increase the number of lonely plans too much, this patch
also changes the multipliers used to be exponential (1, 2, 4, 8, etc.)
rather than linear (1, 2, 3, 4, etc.).

Split the widthHeightOutputDepthInputDepth strategy into two parts - one
for depthwise and one for non-depthwise.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>