Improve estimation accuracy
When estimating the cycle for a Combination, the code balances between
MCE cycles and DRAM bandwidth, taking into account which things can be
done in parallel. However this was being done for the aggregated stats
for the entire network rather than for each pass, which means that the
tradeoff between MCE cycles and DRAM bandwidth is dependent between
passes, which is not correct.

This patch improves the accuracy by performing the calculation for each
pass independently and summing the results. This improves the accuracy
of the estimation and thus allows the compiler to find more performant
Combinations.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>