Fix reshape end-to-end
Change plans for reshape from "no-op" plans with a single DRAM buffer of
the destination shape, to plans which start and end with separate DRAM
buffers, with the input and output shape respectively. These are
connected by DMAs through SRAM.

This avoids issues (mostly in the combiner) with plans that have the
same buffer as both an input and an output, and makes reshape less of a
special case. Specifically, this was causing problems with the buffer
merging, as it resulted in recursive merging which was hard to deal
with.

The new plans are less efficient as they require copying into and out of
SRAM which is unnecessary, but this should not be a big problem as we
don't have many reshapes in our supported networks. If performance
turns out to be an issue, the intention is that these superfluous
copies can be eliminated as a post-process optimisation of the OpGraph,
which should be quite simple.

Simplify ArePlansCompatible as it no longer needs to account for DRAM
reinterprets.

Fix issues with conversion glues which were incorrectly rounding up the
tensor shape and also not setting important SRAM buffer properties
(offset, num stripes and slot size).

When a reshape is at the end of a network, the DRAM buffer in the
reshape plan gets merged with the DRAM buffer in the output plan. The
existing code always choose the first buffer to keep, which in this
case is the reshape buffer. This is wrong because the output buffer is
special and needs to be kept. It is not simple to swap this round and
keep the output buffer, because that gets added to the merged OpGraph
later on. Instead, the fix for this is to create an entirely new buffer
as the result of merging the two DRAM buffers, and add this to the
ending glue of the reshape plan. We always make sure to preserve an
output (or input) buffer when creating this new merged buffer.

When creating the merged OpGraph, we must now account for replacement
buffers in both the stating and ending glues (previously we only
checked for replacement buffers in the starting glue). The ending glue
must also be added to the OpGraph earlier than it used to be, because
we might need to make connections to it due to replaced buffers in the
ending glue, so this code has been moved up.

Fix buffer type in EstimateOnlyPart

Fix tensor shape used in estimation for conversion passes

Fix quantisation info used for input DRAM buffer in FullyConnectedPart

Improve visualisation of replaced buffers

Fix MockParts used in tests to have input/output mappings which are
consistent with their use

Update tests to include a replacement buffer in an ending glue, and
remove the chain of merged buffers, which is no longer possible/necessary.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>