Remove concept of lifetime
Rather than classifying each Op as Cascade or Atomic, use the buffer
stripe and tensor shapes to determine when SRAM can be deallocated.

The lifetime concept was confusing and complicated the problem of
figuring out when to deallocate buffers. This was causing a bug where
buffers were being deallocated too soon in a strategy 1 MCE and PLE
fused cascade.

Instead we simply check if a buffer contains the full tensor, and at
that point we know that all previous buffers are finished with.

Remove a bunch of old combiner checks which are no longer required with
the on-demand plan generation. As well as overcomplicated the code,
these were preventing strategy 1 MCE and PLE plans from being fused
(hiding the above bug and hampering performance).

Merge the PleOperations and SramAllocator parameters passed around a lot
together into a new SectionContext struct, into which a new list of
allocated buffers is stored, so that we know which can be freed.

Rewrite the SRAM deallocation to use stripe shapes rather than
lifetimes, and use a list of allocated buffers rather than just looking
at the buffers in the previous plan. This allows it to deallocate
buffers from anywhere further up in the section, which is needed for
strategy 1 fusing.

Store buffer debug tags in the SramAllocator to help debugging.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>