Fix issue with preimporting in the backend
There was an issue where the EthosNImportTensorhandle wasn't being
selected when preimporting tensor handles.

Refactor the import tensor handles and normal tensor handles to use a
common interface "EthosNBaseTensorHandle". Cast to this base interface
so we don't have to figure out the correct class to use. It will have
already been chosen before.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>