Fix multiple blocks when ifm splitting
The hardware can only support computing one block at a time when
splitting IFMs. This was not being honoured in the plan generation when
the new IFM depth multiplier is greater than one. This is because the
code was assuming that the IFM depth was one block which is no longer
always the case.

This patch fixes this by setting the output depth to one block always.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>