Add Agent IDs to dot file dumps
Add CompiledOpGraph struct to represent an OpGraph that includes details
of command stream generation, currently just the Agent IDs associated
with each Op.

Add SaveCompiledOpGraphToDot function to save this to a dot file,
showing the agent IDs for each Op as well as the grouped agent IDs for
each pass.

Replace DebuggingContext methods with generic method to avoid having to
add a new method here for every corresponding method in
Visualisation.hpp.

This is all to aid debugging.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>