Fix wrong PLE kernel being used
The support library was mapping 32x8 block size to 8x32, leading to the
wrong PLE code being executed. This led to hangs on the firmware.

This has been a confusing area for a while as different parts of the
codebase use different conventions for WxH vs HxW. This patch aims to
remove this confusion by always using WxH.

Also fix plan generation issue seen on the same test with limiting
stripe range leading to missing plans.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>