Add support for split with multiples of 8 channels along axis 3
Until before this patch, split op that splits an input along axis 3
(channel dimension) was supported only in the case when the channel
dimension for all output tensors was a multiple of 16.

Now, with this patch, the support for split layer along axis 3 has
been extended for those cases where all output tensors resulting from
a split layer have channel dimension in multiples of 8.

It has to be noted that this support has been extended to only the
experimental compiler and not the legacy compiler.

Signed-off-by: Shubham <shubham@arm.com>