Use separate environment for driver library tests
When the driver library tests are built, a different Linux kernel
include path is used than the one for the driver library. This is setup
in the driver library tests SConscript by prepending the Linux kernel
include path to use to the CPPPATH variable in the build environment.
However, as the driver library and driver library tests share the same
build environment, the change to the CPPPATH variable affects both of
them causing the driver library to use the wrong Linux kernel include
path.

To ensure that the changes to the build environment for the driver
library tests doesn't affect the driver library build, the driver
library tests will now use a separate build environment by cloning the
one used from the driver library.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>