Fix dangling reference in MockPart
The lifetime of the temporary HardwareCapabilities object passed to the
BasePart constructor in MockPart is only extended until the constructor
exits leaving the BasePart with a dangling reference.

To ensure the HardwareCapabilities object remains valid, the MockPart
will now store a copy of it that the BasePart can be given a reference
to.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>