Add kernel support for TF-A to setup NPU streams
When the kernel module is built for an NPU running in secure mode, it
will no longer setup what SMMU streams the NPU shall use. This is now
taken care of by the TF-A SiP service.

This requires a new SiP service version that supports setting up the
SMMU streams and parsing the new device tree layout introduced with the
multiple SMMU stream support.

The kernel module currently always uses the SMMU streams from the first
available asset allocator so the calls to the SiP service have been
hardcoded to do the same.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>