Remove superfluous dependencies with branching
Previously the code was adding multiple schedule dependencies onto the
same OfmS agent, when the output buffer of the OfmS was consumed by
multiple IfmS. This was incorrect, because we don't support an
unbounded number of schedule dependencies, and also this doesn't
correctly describe the linearised command stream.

This patch tweaks the dependencies generated so that when there is more
than one consumer of an OfmS, the dependencies are linearised so that
the IfmS of the second branch depends on the final OfmS of the previous
branch.

This has been implemented by checking for "fence points" in the command
stream, where the combiner will have dellocated all previous SRAM
buffers and so we need to ensure that no agents further down can write
anything into SRAM until the agents before the fence point has
finished.

This revealed a couple of missing dependencies (e.g. for PleL) which
have now been added too.

This simplifies some existing of the existing dependency generation by
making it more generic, and avoiding the need to look up in the
OpGraph.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>