Fix calculation of ifmDeltaDefault for strided
The previous calculation was incorrect because it was mixing up stripe
shapes and tensor shapes, resulting in very large delta values which
were both incorrect and also overflowed the maximum value supported by
the hardware.

Replace the calculation with the equivalent of the one we do for the
non-strided case.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>