Add support for concat with multiples of 8 channels along axis 3
Until before this patch, concat op that concats inputs along axis 3
(channel dimension) was supported only in the case when the channel
dimension for all input tensors was a multiple of 16.

Now, with this patch, the support for concat layer along axis 3 has
been extended for those cases where all input tensors have channel
dimension in multiple of 8.

Please note that this extened support is only provided for the case
when the experimental compiler is enabled.

Signed-off-by: Shubham <shubham@arm.com>