Move DataType field to Buffer
Previously this was being stored on MceOp and PleOp, and this patch
removes it from there and adds it to the Buffer type. This is more
consistent with other similar properties like tensor shape and
quantisation info.

It also allows correctly setting the fcafInfo.signedActivation flag as
the information we need is now available.

The intermediate buffer dumping now uses the correct format in the dump
filename, allowing these dump files to be correctly interpreted by any
automatic tools.

Change to use support library's DataType enum rather than the command
stream's enum, as this is more appropriate (and requires less
conversions).

Signed-off-by: Rob Hughes <robert.hughes@arm.com>