Add intermediate buffer size to CompiledNetwork
In upcoming changes to the kernel module and driver library, support
will be added to allow the user to import a buffer to hold a network's
intermediate data.

To facilitate this, the CompiledNetwork class has been changed to include
a function to specify how large of a buffer the network needs for its
intermediate data.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>