Fix non-cascading using output buffer as input
Previously, there hasn't been any restrictions for what type of buffer
that could be used as the input for a node. This allowed the
non-cascading compiler to convert an intermediate buffer to an output
buffer for an OutputNode and still use the buffer as an input to another
node.

However, with recent changes to restrict how the different buffer types
can be used, it is no longer allowed to use an output buffer as an input
to another node. Therefore, the non-cascading compiler has been changed
to copy the intermediate buffer rather than converting it to an output
buffer, if the intermediate buffer is used by another node.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>