Fix uninitialized array usage in XmlToBinary tests
The convertCharsToArray function returns an array that has been default
initialized, which means that any position in the array that isn't set
to a value will be filled with some indeterminate data.

In this case, it doesn't cause any issues because the array is used to
store a null-terminated string but to make debugging possible issues
easier the array is now being zero initialized.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>