Fix parameter shadowing in ReshapePart::GetPlans
The GetPlans function in ReshapePart declares local variables with the
same name as the parameters to the function. This means the parameters
are being shadowed by the local variables.

In this case, the function doesn't use the shadowed parameters so
instead of renaming the local variables, the parameter names have been
removed from the function.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>