Fix order of agents with addition IFM streamers
The code that was sorting the OpGraph was causing one of the IFM
streamers for the PleS to be in the wrong order (run much earlier than
it should have been).

This sorting is not necessary as the OpGraph is already in execution
order, and removing this code fixes the IFM streamer order.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>