Generate more plans for Split
Remove SplitOp (including associated command stream generation and
estimation code) and instead use DmaOp. This allows
SplitPart to generate multiple plans, each with different stripe
shapes. This was not possible before because SplitOp does not have a
stripe shape, as this was decided later in the command stream
generation. This is also a simplification because we can re-use all the
existing DmaOp code.

The combiner can then choose the best plan (best stripe shape), leading
to much improved performance for larger splits.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>