Fix wrong size in carveout sub allocator clean up
When a carveout sub allocator is destroyed, the memory holding the sub
allocator should be zeroed out before being freed to not leave any
information behind. The zeroing of the memory is currently using the
wrong struct type to get the size so not all the memory is being
cleared.

This has now been fixed so the correct struct type is used when getting
the size to zero out the memory.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>