Include subgraph index in layer name
Set the name of the PreCompiledLayer that the Arm NN backend creates to
include the subgraph index. This makes it easier to match up the
subgraph logs and dumps with the Arm NN dot file dump.

Also remove the static "instance ID" counter from
EthosNSubgraphViewConverter and instead have a non-static member in the
backend object.

This means subgraph indices will restart from zero for each network,
reduces global state, and simplifies the unit tests.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>