Add Logger functions that use va_list format args
In some cases a variadic function has already been used to setup the
information that should be logged so only a va_list is available to be
passed on to the Logger functions as the format args. However, there are
currently only variadic Logger functions available and passing a va_list
to one to those functions will result in undefined behavior because the
function will create a va_list of the passed va_list and then go beyond
the va_list when retrieving the values for the format string.

To solve this, a new set of Logger functions that accepts the format
arguments as a va_list have been added.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>