Add IOCTL API to create process memory allocator
As part of per process memory protection support, a new IOCTL API is
added to create a process memory allocator that processes will use to
allocate memory for its assets. The old IOCTL APIs for buffer
create/import and register network are deprecated. The old APIs still
work but they will be removed in the future. With the new IOCTL APIs,
the userspace process requests a memory allocator via
ETHOSN_IOCTL_CREATE_PROC_MEM_ALLOCATOR. This reserves a free asset
allocator and returns a fd which is used as a reference to create/import
buffers and register networks. This enforces a one asset allocator per
process limit and also ensures that all the memory assets created using
the asset allocator are freed before the asset allocator itself is
freed, so the asset allocator isn't reused while the memory assets made
using it still exist.

Currently there is only one asset allocator available for use. Due to
this only one process at a time can create a process memory allocator
using the new IOCTL APIs. This limitation will be removed in the future.

Signed-off-by: Rajasekaran Kalidoss <rajasekaran.kalidoss@arm.com>