Prefer NHWC concat format in some cases
We prefer to use NHWCB for concat if possible, but this isn't always the
best choice. This patch attempts to make a slightly better decision, by
checking if all inputs to the concat are going to be NHWC format, and
if so prefers NHWC format in order to minimise the number of format
conversions required.

This speeds up performance for some networks.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>