Account for DMA chunking in perf estimation
Even if two strategies have the same number of stripes, if one of them
requires those stripes to be split into chunks in order to be DMA'd,
then this will lead to worse performance. Up until now, the performance
estimator in the compiler did not account for this.

This patch addresses this by including the number of chunks in the
calculation for the number of stripes.

This results in better plans being chosen, and better performance.

Because of this, AddTempGlues in the Combiner needed to be updated to
select an appropriate DRAM format, rather than the old behaviour which
was just to assume NHWCB. This is because NHWCB can now be evaluated to
give worse performance than FCAF because of chunking and this might
lead to suboptimal plans being chosen. Note that this format won't
necessarily be the same as the format used in the final compilation,
because we don't know what other users of this buffer will require.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>