Account for non-parallel output DMAs
Previously the code in GetOutputStats assumed that the output could be
copied in parallel with other processing, but this is only true if we
have enough space in the output tile for more than one buffer. This led
to no advantage in the performance estimation for double-buffered
plans.

This patch makes the code account for double-buffering, which means that
more optimal plans are chosen, resulting in increased performance.

GetOutputStats has been split into Legacy and Cascading versions, so as
not to affect the behaviour of the legacy performance estimation.

Also remove unused GetInputStats

Signed-off-by: Rob Hughes <robert.hughes@arm.com>