Fix weight tensor depth for striding
The depth of the weight tensor in the OpGraph was recently updated in
patch "Fix weight buffer tensor shape" to account for striding, but the
fix wasn't quite right as the simple multiplication didn't account for
cases where the original weight channels isn't an exact multiple of the
product of the strides. In this case a more complicated calculation is
required, which we already have a utility function for.

Therefore this patch replaces the simple multiplication with a call to
the existing utility function.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>