Generate more lonely plans
Previously, when generating lonely plans we first tried to generate
beginning plans, and only if none of these were valid did we try to
generate actual lonely plans. This was done to make sure we chose plans
that minimise the effects of stripe overhead.

Since that was done however, the estimation code was updated to take
into account stripe overhead directly, and so we can go back to
generating all the lonely plans and let the estimation code decide
which is best, taking into account stripe overhead. This gives us more
choice of plans, and in some cases this means we can pick better ones.

Fix unit test expected results.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>