Fix outer ratios on IfmS and MceS for depthwise
The calculation for the outer ratio was assuming that the only case
where a non-trivial outer ratio is needed was when height and depth
splitting("strategy 6"). This wasn't correct, as we also need a
non-trivial outer ratio when doing other kinds of strategies,
particularly for depthwise where we don't need to split IFMs, and need
to correctly include boundary data at the start of each row/column.

This patch expands the logic to set the outer ratio to include these
cases that were previously missed.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>