Pad output channels of Concat to improve performance
When concating along the channels dimension and the sizes of the input
tensors (along the channels dimension) are not a multiple of 16, then
we are forced to choose a stripe shape of 8x8 (width x height) so that
the DMAs are supported by the firmware. This can lead to a very large
number of stripes if the tensor being split has large XY dimensions,
which can lead to poor performance.

This patch allows the support library to "pad" the output channels of
the Concat layer such that the subtensors are aligned on boundaries of
16. This allows larger stripe shapes to be used, and thus there are
less stripes for the firmware to process and performance increases.

This method also lifts the restriction on the Concat sizes being
multiples of 8 - we can now support any concat size on the experimental
compiler.

The ConcatPart now takes a separate array of offsets, as these are no
longer necessarily the same as the input tensor shapes (due to
padding).

Signed-off-by: Rob Hughes <robert.hughes@arm.com>