Account for wraparound in profiling command idx
For larger command streams, the number of commands executed can be
larger than the profiling field has space for (currently 10 bits => max
value of 1023). This leads to profiling results being difficult to
interpret because the command number appears to restart from zero.

This patch adds a heuristic in the driver library when dumping profiling
information so that wraparound is detected and the original command
indices are restored.

In order to do this, the driver library dumping code now extracts
individual fields from the firmware profiling entries and writes those
fields to the JSON, rather than writing the binary value directly. This
also simplifies further processing of profiling dumps.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>