Update README.md
The README.md file has been updated with a link to the patch that must
be applied to add support for SMMU streams. Also correct the level for
the 'Build the Ethos-N NPU driver' section.  Moreover, update the
'Download the Ethos-N driver stack' section to point to the correct
branch

Signed-off-by: Mohamed Elzahhar <Mohamed.Elzahhar@arm.com>