Fix upper bounds of stripe shape loops
There are several places in the support library where we loop over
possible stripe shapes. Sometimes we should be including stripe shapes
that cover the full tensor, and sometimes we shouldn't be. We also use
a logarithmic spacing of stripe shapes to reduce the compilation time.
All this leads to confusing and duplicated code, and has some bugs
where we get the upper bound wrong, leading to missing or extra plans.

This patch centralises this logic into a new StripeShapeLoop helper
object, which handles the looping. Existing uses have been changed to
use this new helper, in the process fixing several unintended
behaviours.

The main benefit of this patch is the improved readability of the code,
but also yields some performance benefits.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>