Pad input channels of Split to improve performance
When splitting along the channels dimension and the sizes of the output
tensors (along the channels dimension) are not a multiple of 16, then
we are forced to choose a stripe shape of 8x8 (width x height) so that
the DMAs are supported by the firmware. This can lead to a very large
number of stripes if the tensor being split has large XY dimensions,
which can lead to poor performance.

This patch allows the support library to "pad" the input channels of the
Split layer such that the subtensors are aligned on boundaries of 16.
This allows larger stripe shapes to be used, and thus there are less
stripes for the firmware to process and performance increases.

This method also lifts the restriction on the split sizes being
multiples of 8 - we can now support any split size on the experimental
compiler.

This means that EstimateOnly is no longer a valid case for Split with
the experimental compiler and so this code has been removed.

The SplitPart now takes a separate array of offsets, as these are no
longer necessarily the same as the output tensor shapes (due to
padding).

Add an overload of the SupportQueries constructor (and friends) to allow
enabling the experimental compiler without having to use the
environment variable. This is necessary for unit testing of cascading
code.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>