Prevent kernel module locking up
When the kernel module is receiving lots of messages from the firmware,
e.g. when receiving log messages, it can appear to lock up while trying
to process all the messages.

This patch introduces a limit to the number of messages processed in
each work queue handler function call, so that there is an opportunity
to interrupt this (e.g. to stop the inference).

This can save a lot of time when debugging.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>