Set ethosn_dma_info pointers to NULL once freed
The allocator functions were freeing the ethosn_dma_info struct data but
not setting the pointer to NULL, which meant that we would be left with
dangling pointers that make it difficult to know if the pointer is
valid or not, and could cause hard-to-detect bugs.

This patch updates the allocators to set the pointer to NULL after
deallocating it.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>