Fix kernel module crash when firmware size changes
If the firmware binary size increases while the kernel module is loaded,
and the kernel module tries to reload the firmware (e.g. due to runtime
power management), then it will attempt to copy the larger firmware
over a buffer allocated for the original (smaller) size, and crash.

This patch fixes this by re-allocating the buffer if necessary.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>